exit

+++ integration-tests/jail/run.sh
@@ -6,6 +6,10 @@ echo "docker compose ok"
 #echo "done"
 pip3 install docker
 python3 ./disk/jail_test.py
+ret=$?
+if [ $ret -ne 0 ]; then
+    exit -1
+fi
 echo "test finished"
 sleep 4
 docker-compose down
